+++ hackathon-flink-scala/src/main/scala/eu/streamline/hackathon/flink/scala/job/FlinkScalaJob.scala
@@ -25,74 +25,37 @@ object FlinkScalaJob {
   case class TempRule(name: String, threshold: Integer)
 
   def main(args: Array[String]): Unit = {
+    val parameters = ParameterTool.fromArgs(args)
+    val pathToGDELT = parameters.get("path")
+    val pathToWeather = parameters.get("wpath")
+    val startWebUI = parameters.getBoolean("webui", false)
+    val config = new Configuration()
+    config.setBoolean(ConfigConstants.LOCAL_START_WEBSERVER, true)
+    val env = if (startWebUI) StreamExecutionEnvironment.createLocalEnvironmentWithWebUI(config)
+      else StreamExecutionEnvironment.getExecutionEnvironment
+    env.setStreamTimeCharacteristic(TimeCharacteristic.EventTime)
+    env.setStateBackend(new RocksDBStateBackend("file:///tmp/flinkrocksdb", true))
+
 
     // join events to weather data on the same day
     val joinWindowDays = 1
 
     // we can use tumbling or sliding windows here
     val avgToneWindow = TumblingEventTimeWindows.of(Time.days(1))
-      // SlidingEventTimeWindows.of(Time.days(4), Time.days(2))
-
-    val parameters = ParameterTool.fromArgs(args)
-    val pathToGDELT = parameters.get("path")
-    val pathToWeather = parameters.get("wpath")
-    val startWebUI = parameters.getBoolean("webui", false)
-
-    val config = new Configuration()
-    config.setBoolean(ConfigConstants.LOCAL_START_WEBSERVER, true)
-    val env = if (startWebUI) StreamExecutionEnvironment.createLocalEnvironmentWithWebUI(config)
-      else StreamExecutionEnvironment.getExecutionEnvironment
-
-    env.setStreamTimeCharacteristic(TimeCharacteristic.EventTime)
-    env.setStateBackend(new RocksDBStateBackend("file:///tmp/flinkrocksdb", true))
+    // SlidingEventTimeWindows.of(Time.days(4), Time.days(2))
 
     val tempRules = Array(TempRule("low", 5), TempRule("mid", 25), TempRule("high", Integer.MAX_VALUE))
-
-    val weatherRawStream = env.readTextFile(pathToWeather).setParallelism(1).name("Weather Source")
-
-
-    val dateFormat = new SimpleDateFormat("yyyyMMdd")
-    val weatherStream = weatherRawStream
-      .mapWith(line => {
-        val elements = line.split(",")
-        val elementTuple = (elements(0), dateFormat.parse(elements(1)), elements(2), elements(3))
-        WeatherRecord.tupled(elementTuple)
-      })
-      .filter(weatherRecord => weatherRecord.date.after(dateFormat.parse("20170131")))
-      .assignTimestampsAndWatermarks(new BoundedOutOfOrdernessTimestampExtractor[WeatherRecord](Time.seconds(0)) {
-          override def extractTimestamp(element: WeatherRecord): Long = {
-            element.date.getTime
-          }
-      })
-
+    val weatherStream = Streams.getWeatherStream(env, pathToWeather)
     val maxTempWeather = weatherStream.filterWith{element => element.element == "TMAX"}
 
-    // TODO remove; these typeinfos seem to be not used?
-//    implicit val typeInfo = createTypeInformation[GDELTEvent]
-//    implicit val dateInfo = createTypeInformation[Date]
+    val gdeltEventStream = Streams.getGDELTEvents(env, pathToGDELT)
 
-    val gdeltEventStream = env
-      .readFile[GDELTEvent](new GDELTInputFormat(new Path(pathToGDELT)), pathToGDELT)
-      .setParallelism(1)
-      .name("GDELT Event Source")
-      .filter((event: GDELTEvent) => {
-        event.actor1Code_countryCode != null &
-        // we only have weather data for USA anyway
-        event.actor1Code_countryCode == "USA" &
-        event.eventGeo_lat != null &
-        event.eventGeo_long != null &
-        event.isRoot //& event.eventRootCode.equals("14")
-    }).assignTimestampsAndWatermarks(new BoundedOutOfOrdernessTimestampExtractor[GDELTEvent](Time.seconds(0)) {
-      override def extractTimestamp(element: GDELTEvent): Long = {
-        element.day.getTime
-      }
-    })
 
     val stationLocator = new StationLocator()
-
     val eventNearestStations = gdeltEventStream.map(event => {
       (event, stationLocator.nearest(event))
     }).name("Station Locator")
+
 
     case class EventWeather(id: Integer, date: Instant, tone: Double, temp: Double)
     // classified event base on temp rules
